[REFACTOR] 커스텀 필드
